--- a/.changeset/modern-numbers-impress.md
+++ b/.changeset/modern-numbers-impress.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `selector-anb-no-unmatchable` performance