--- a/lib/rules/selector-anb-no-unmatchable/index.js
+++ b/lib/rules/selector-anb-no-unmatchable/index.js
@@ -10,6 +10,7 @@ const isStandardSyntaxRule = require('../../utils/isStandardSyntaxRule');
 const report = require('../../utils/report');
 const ruleMessages = require('../../utils/ruleMessages');
 const validateOptions = require('../../utils/validateOptions');
+const hasANPlusBNotationPseudoClasses = require('../../utils/hasANPlusBNotationPseudoClasses');
 
 const ruleName = 'selector-anb-no-unmatchable';
 
@@ -50,6 +51,8 @@ const rule = (primary) => {
 			}
 
 			ruleNode.selectors.forEach((selector) => {
+				if (!hasANPlusBNotationPseudoClasses(selector)) return;
+
 				let cssTreeSelector;
 
 				try {