--- a/lib/utils/hasANPlusBNotationPseudoClasses.js
+++ b/lib/utils/hasANPlusBNotationPseudoClasses.js
@@ -5,13 +5,11 @@ const {
 	aNPlusBOfSNotationPseudoClasses,
 } = require('../reference/selectors');
 
-const HAS_A_N_PLUS_B_NOTATION_PSEUDO_CLASSES = new RegExp(
-	`\\b:(?:${[
-		...aNPlusBNotationPseudoClasses.values(),
-		...aNPlusBOfSNotationPseudoClasses.values(),
-	].join('|')})\\(`,
-	'i',
-);
+const classes = [
+	...aNPlusBNotationPseudoClasses.values(),
+	...aNPlusBOfSNotationPseudoClasses.values(),
+].join('|');
+const HAS_A_N_PLUS_B_NOTATION_PSEUDO_CLASSES = new RegExp(`\\b:(?:${classes})\\(`, 'i');
 
 /**
  * Check if a selector contains any pseudo class function that might contain an An+B notation