<!-- Each pull request must be associated with an open issue unless it's a documentation fix. If a corresponding issue does not exist, please create one so we can discuss the change first. -->

<!-- Please answer the following. We close pull requests that don't. -->

> Which issue, if any, is this issue related to?

See : https://github.com/stylelint/stylelint/issues/6869

> Is there anything in the PR that needs further explanation?

No, it's self-explanatory.

-----

before : 

```
Warnings: 0
Mean: 171.01108333333332 ms
Deviation: 11.09835497913841 ms
```

after :

```
Warnings: 0
Mean: 115.6843908125 ms
Deviation: 7.734276833118593 ms
```